Work Completed on 1-24-2019
Work Completed on 1-24-2019 - Changed the wpuserid column in the Users table from smallint(6) to bigint(255) to accomodate the large WP User IDs on wordpress.com websites.

+++ includes/class-wpbooklist-general-functions.php
@@ -1031,7 +1031,7 @@ if ( ! class_exists( 'WPBookList_General_Functions', false ) ) :
 			            firstname varchar(190),
 			            lastname varchar(255),
 			            datecreated varchar(255),
-			            wpuserid smallint(6),
+			            wpuserid bigint(255),
 			            email varchar(255),
 			            username varchar(255),
 			            role varchar(255),
